chore(operator): fixed pre- and post-deployment
