+++ examples/podtatohead-deployment/app.yaml
@@ -4,7 +4,7 @@ metadata:
   name: podtato-head
   namespace: podtato-kubectl
 spec:
-  version: "1.0"
+  version: "1.3"
   workloads:
     - name: podtato-head-left-arm
       version: 0.1.0
@@ -16,5 +16,7 @@ spec:
       version: 0.1.0
     - name: podtato-head-left-arm
       version: 0.1.0
-    - name: podtato-head-hats
+    - name: podtato-head-hat
       version: 0.1.0
+  postDeploymentTasks:
+    - post-deployment-hello
