+++ operator/config/crd/bases/lifecycle.keptn.sh_keptnappversions.yaml
@@ -12,24 +12,20 @@ spec:
     kind: KeptnAppVersion
     listKind: KeptnAppVersionList
     plural: keptnappversions
+    shortNames:
+    - kav
     singular: keptnappversion
   scope: Namespaced
   versions:
   - additionalPrinterColumns:
-    - jsonPath: .spec.app
+    - jsonPath: .spec.appname
       name: AppName
-      type: string
-    - jsonPath: .spec.workloadName
-      name: Workload
       type: string
     - jsonPath: .spec.version
       name: Version
       type: string
     - jsonPath: .status.preDeploymentStatus
       name: PreDeploymentStatus
-      type: string
-    - jsonPath: .status.workloadStatus
-      name: WorkloadStatus
       type: string
     - jsonPath: .status.workloadOverallStatus
       name: WorkloadOverallStatus
