+++ operator/controllers/keptnappversion/controller.go
@@ -32,10 +32,7 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/propagation"
 	"k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
-	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	"k8s.io/apimachinery/pkg/runtime"
@@ -106,7 +103,7 @@ func (r *KeptnAppVersionReconciler) Reconcile(ctx context.Context, req ctrl.Requ
 			r.recordEvent(phase, "Warning", appVersion, "Failed", "has failed")
 			return ctrl.Result{Requeue: true, RequeueAfter: 60 * time.Second}, nil
 		}
-		state, err := r.reconcilePreDeployment(ctx, appVersion)
+		state, err := r.reconcilePrePostDeployment(ctx, appVersion, common.PreDeploymentCheckType)
 		if err != nil {
 			span.SetStatus(codes.Error, err.Error())
 			r.recordEvent(phase, "Warning", appVersion, "ReconcileErrored", "could not get reconciled")
@@ -149,7 +146,7 @@ func (r *KeptnAppVersionReconciler) Reconcile(ctx context.Context, req ctrl.Requ
 			r.recordEvent(phase, "Warning", appVersion, "Failed", "has failed")
 			return ctrl.Result{Requeue: true, RequeueAfter: 60 * time.Second}, nil
 		}
-		state, err := r.reconcilePostDeployment(ctx, appVersion)
+		state, err := r.reconcilePrePostDeployment(ctx, appVersion, common.PostDeploymentCheckType)
 		if err != nil {
 			span.SetStatus(codes.Error, err.Error())
 			r.recordEvent(phase, "Warning", appVersion, "ReconcileErrored", "could not get reconciled")
@@ -203,149 +200,6 @@ func (r *KeptnAppVersionReconciler) SetupWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&klcv1alpha1.KeptnAppVersion{}, builder.WithPredicates(predicate.GenerationChangedPredicate{})).
 		Complete(r)
-}
-
-func (r *KeptnAppVersionReconciler) createKeptnTask(ctx context.Context, namespace string, appVersion *klcv1alpha1.KeptnAppVersion, taskDefinition string, checkType common.CheckType) (string, error) {
-
-	ctx, span := r.Tracer.Start(ctx, "create_app_task", trace.WithSpanKind(trace.SpanKindProducer))
-	defer span.End()
-
-	semconv.AddAttributeFromAppVersion(span, *appVersion)
-
-	// create TraceContext
-	// follow up with a Keptn propagator that JSON-encoded the OTel map into our own key
-	traceContextCarrier := propagation.MapCarrier{}
-	otel.GetTextMapPropagator().Inject(ctx, traceContextCarrier)
-
-	phase := common.KeptnPhaseType{
-		ShortName: "KeptnTaskCreate",
-		LongName:  "Keptn Task Create",
-	}
-
-	newTask := &klcv1alpha1.KeptnTask{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        common.GenerateTaskName(checkType, taskDefinition),
-			Namespace:   namespace,
-			Annotations: traceContextCarrier,
-		},
-		Spec: klcv1alpha1.KeptnTaskSpec{
-			Version:          appVersion.Spec.Version,
-			AppName:          appVersion.Spec.AppName,
-			TaskDefinition:   taskDefinition,
-			Parameters:       klcv1alpha1.TaskParameters{},
-			SecureParameters: klcv1alpha1.SecureParameters{},
-			Type:             checkType,
-		},
-	}
-	err := controllerutil.SetControllerReference(appVersion, newTask, r.Scheme)
-	if err != nil {
-		r.Log.Error(err, "could not set controller reference:")
-	}
-	err = r.Client.Create(ctx, newTask)
-	if err != nil {
-		r.Log.Error(err, "could not create KeptnTask")
-		r.recordEvent(phase, "Warning", appVersion, "CreateFailed", "could not create KeptnTask")
-		return "", err
-	}
-	r.recordEvent(phase, "Normal", appVersion, "Created", "created")
-
-	return newTask.Name, nil
-}
-
-func (r *KeptnAppVersionReconciler) reconcileTasks(ctx context.Context, checkType common.CheckType, appVersion *klcv1alpha1.KeptnAppVersion) ([]klcv1alpha1.TaskStatus, common.StatusSummary, error) {
-	phase := common.KeptnPhaseType{
-		ShortName: "ReconcileTasks",
-		LongName:  "Reconcile Tasks",
-	}
-
-	var tasks []string
-	var statuses []klcv1alpha1.TaskStatus
-
-	switch checkType {
-	case common.PreDeploymentCheckType:
-		tasks = appVersion.Spec.PreDeploymentTasks
-		statuses = appVersion.Status.PreDeploymentTaskStatus
-	case common.PostDeploymentCheckType:
-		tasks = appVersion.Spec.PostDeploymentTasks
-		statuses = appVersion.Status.PostDeploymentTaskStatus
-	}
-
-	var summary common.StatusSummary
-	summary.Total = len(tasks)
-	// Check current state of the PrePostDeploymentTasks
-	var newStatus []klcv1alpha1.TaskStatus
-	for _, taskDefinitionName := range tasks {
-		var oldstatus common.KeptnState
-		for _, ts := range statuses {
-			if ts.TaskDefinitionName == taskDefinitionName {
-				oldstatus = ts.Status
-			}
-		}
-
-		taskStatus := GetTaskStatus(taskDefinitionName, statuses)
-		task := &klcv1alpha1.KeptnTask{}
-		taskExists := false
-
-		if oldstatus != taskStatus.Status {
-			r.recordEvent(phase, "Normal", appVersion, "TaskStatusChanged", fmt.Sprintf("task status changed from %s to %s", oldstatus, taskStatus.Status))
-		}
-
-		// Check if task has already succeeded or failed
-		if taskStatus.Status == common.StateSucceeded || taskStatus.Status == common.StateFailed {
-			newStatus = append(newStatus, taskStatus)
-			continue
-		}
-
-		// Check if Task is already created
-		if taskStatus.TaskName != "" {
-			err := r.Client.Get(ctx, types.NamespacedName{Name: taskStatus.TaskName, Namespace: appVersion.Namespace}, task)
-			if err != nil && errors.IsNotFound(err) {
-				taskStatus.TaskName = ""
-			} else if err != nil {
-				return nil, summary, err
-			}
-			taskExists = true
-		}
-
-		// Create new Task if it does not exist
-		if !taskExists {
-			taskName, err := r.createKeptnTask(ctx, appVersion.Namespace, appVersion, taskDefinitionName, checkType)
-			if err != nil {
-				return nil, summary, err
-			}
-			taskStatus.TaskName = taskName
-			taskStatus.SetStartTime()
-		} else {
-			// Update state of Task if it is already created
-			taskStatus.Status = task.Status.Status
-			if taskStatus.Status.IsCompleted() {
-				taskStatus.SetEndTime()
-			}
-		}
-		// Update state of the Check
-		newStatus = append(newStatus, taskStatus)
-	}
-
-	for _, ns := range newStatus {
-		summary = common.UpdateStatusSummary(ns.Status, summary)
-	}
-	if common.GetOverallState(summary) != common.StateSucceeded {
-		r.recordEvent(phase, "Warning", appVersion, "NotFinished", "has not finished")
-	}
-	return newStatus, summary, nil
-}
-
-func GetTaskStatus(taskName string, instanceStatus []klcv1alpha1.TaskStatus) klcv1alpha1.TaskStatus {
-	for _, status := range instanceStatus {
-		if status.TaskDefinitionName == taskName {
-			return status
-		}
-	}
-	return klcv1alpha1.TaskStatus{
-		TaskDefinitionName: taskName,
-		Status:             common.StatePending,
-		TaskName:           "",
-	}
 }
 
 func (r *KeptnAppVersionReconciler) recordEvent(phase common.KeptnPhaseType, eventType string, appVersion *klcv1alpha1.KeptnAppVersion, shortReason string, longReason string) {
