+++ operator/controllers/keptnworkloadinstance/reconcile_prepostdeployment.go
@@ -2,26 +2,18 @@ package keptnworkloadinstance
 
 import (
 	"context"
+	"fmt"
 	klcv1alpha1 "github.com/keptn-sandbox/lifecycle-controller/operator/api/v1alpha1"
 	"github.com/keptn-sandbox/lifecycle-controller/operator/api/v1alpha1/common"
+	"github.com/keptn-sandbox/lifecycle-controller/operator/api/v1alpha1/semconv"
+	"go.opentelemetry.io/otel"
+	"go.opentelemetry.io/otel/propagation"
+	"go.opentelemetry.io/otel/trace"
+	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
-
-func (r *KeptnWorkloadInstanceReconciler) reconcilePostDeployment(ctx context.Context, workloadInstance *klcv1alpha1.KeptnWorkloadInstance) (common.KeptnState, error) {
-	newStatus, postDeploymentState, err := r.reconcileChecks(ctx, common.PostDeploymentCheckType, workloadInstance)
-	if err != nil {
-		return common.StateUnknown, err
-	}
-	overallState := common.GetOverallState(postDeploymentState)
-	workloadInstance.Status.PostDeploymentStatus = overallState
-	workloadInstance.Status.PostDeploymentTaskStatus = newStatus
-
-	// Write Status Field
-	err = r.Client.Status().Update(ctx, workloadInstance)
-	if err != nil {
-		return common.StateUnknown, err
-	}
-	return overallState, nil
-}
 
 func (r *KeptnWorkloadInstanceReconciler) reconcilePrePostDeployment(ctx context.Context, workloadInstance *klcv1alpha1.KeptnWorkloadInstance, checkType common.CheckType) (common.KeptnState, error) {
 	newStatus, state, err := r.reconcileTasks(ctx, checkType, workloadInstance)
@@ -45,4 +37,150 @@ func (r *KeptnWorkloadInstanceReconciler) reconcilePrePostDeployment(ctx context
 		return common.StateUnknown, err
 	}
 	return overallState, nil
+}
+
+func (r *KeptnWorkloadInstanceReconciler) getKeptnTask(ctx context.Context, taskName string, namespace string) (*klcv1alpha1.KeptnTask, error) {
+	task := &klcv1alpha1.KeptnTask{}
+	err := r.Client.Get(ctx, types.NamespacedName{Name: taskName, Namespace: namespace}, task)
+	if err != nil {
+		return task, err
+	}
+	return task, nil
+}
+
+func (r *KeptnWorkloadInstanceReconciler) createKeptnTask(ctx context.Context, namespace string, workloadInstance *klcv1alpha1.KeptnWorkloadInstance, taskDefinition string, checkType common.CheckType) (string, error) {
+	ctx, span := r.Tracer.Start(ctx, fmt.Sprintf("create_%s_deployment_task", checkType), trace.WithSpanKind(trace.SpanKindProducer))
+	defer span.End()
+
+	semconv.AddAttributeFromWorkloadInstance(span, *workloadInstance)
+
+	// create TraceContext
+	// follow up with a Keptn propagator that JSON-encoded the OTel map into our own key
+	traceContextCarrier := propagation.MapCarrier{}
+	otel.GetTextMapPropagator().Inject(ctx, traceContextCarrier)
+	newTask := &klcv1alpha1.KeptnTask{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:        common.GenerateTaskName(checkType, taskDefinition),
+			Namespace:   namespace,
+			Annotations: traceContextCarrier,
+		},
+		Spec: klcv1alpha1.KeptnTaskSpec{
+			Version:          workloadInstance.Spec.Version,
+			AppName:          workloadInstance.Spec.AppName,
+			TaskDefinition:   taskDefinition,
+			Parameters:       klcv1alpha1.TaskParameters{},
+			SecureParameters: klcv1alpha1.SecureParameters{},
+			Type:             checkType,
+		},
+	}
+	err := controllerutil.SetControllerReference(workloadInstance, newTask, r.Scheme)
+	if err != nil {
+		r.Log.Error(err, "could not set controller reference:")
+	}
+	err = r.Client.Create(ctx, newTask)
+	if err != nil {
+		r.Log.Error(err, "could not create KeptnTask")
+		r.Recorder.Event(workloadInstance, "Warning", "KeptnTaskNotCreated", fmt.Sprintf("Could not create KeptnTask / Namespace: %s, Name: %s ", newTask.Namespace, newTask.Name))
+		return "", err
+	}
+	r.Recorder.Event(workloadInstance, "Normal", "KeptnTaskCreated", fmt.Sprintf("Created KeptnTask / Namespace: %s, Name: %s ", newTask.Namespace, newTask.Name))
+
+	return newTask.Name, nil
+}
+
+func (r *KeptnWorkloadInstanceReconciler) reconcileTasks(ctx context.Context, checkType common.CheckType, workloadInstance *klcv1alpha1.KeptnWorkloadInstance) ([]klcv1alpha1.TaskStatus, common.StatusSummary, error) {
+	var tasks []string
+	var statuses []klcv1alpha1.TaskStatus
+
+	phase := common.KeptnPhaseType{
+		ShortName: "ReconcileTasks",
+		LongName:  "Reconcile Tasks",
+	}
+
+	switch checkType {
+	case common.PreDeploymentCheckType:
+		tasks = workloadInstance.Spec.PreDeploymentTasks
+		statuses = workloadInstance.Status.PreDeploymentTaskStatus
+	case common.PostDeploymentCheckType:
+		tasks = workloadInstance.Spec.PostDeploymentTasks
+		statuses = workloadInstance.Status.PostDeploymentTaskStatus
+	}
+
+	var summary common.StatusSummary
+	summary.Total = len(tasks)
+	// Check current state of the PrePostDeploymentTasks
+	var newStatus []klcv1alpha1.TaskStatus
+	for _, taskDefinitionName := range tasks {
+		var oldstatus common.KeptnState
+
+		for _, ts := range statuses {
+			if ts.TaskDefinitionName == taskDefinitionName {
+				oldstatus = ts.Status
+			}
+		}
+
+		taskStatus := GetTaskStatus(taskDefinitionName, statuses)
+		task := &klcv1alpha1.KeptnTask{}
+		taskExists := false
+
+		if oldstatus != taskStatus.Status {
+			r.recordEvent(phase, "Normal", workloadInstance, "TaskStatusChanged", fmt.Sprintf("task status changed from %s to %s", oldstatus, taskStatus.Status))
+		}
+
+		// Check if task has already succeeded or failed
+		if taskStatus.Status == common.StateSucceeded || taskStatus.Status == common.StateFailed {
+			newStatus = append(newStatus, taskStatus)
+			continue
+		}
+
+		// Check if Task is already created
+		if taskStatus.TaskName != "" {
+			err := r.Client.Get(ctx, types.NamespacedName{Name: taskStatus.TaskName, Namespace: workloadInstance.Namespace}, task)
+			if err != nil && errors.IsNotFound(err) {
+				taskStatus.TaskName = ""
+			} else if err != nil {
+				return nil, summary, err
+			}
+			taskExists = true
+		}
+
+		// Create new Task if it does not exist
+		if !taskExists {
+			taskName, err := r.createKeptnTask(ctx, workloadInstance.Namespace, workloadInstance, taskDefinitionName, checkType)
+			if err != nil {
+				return nil, summary, err
+			}
+			taskStatus.TaskName = taskName
+			taskStatus.SetStartTime()
+		} else {
+			// Update state of Task if it is already created
+			taskStatus.Status = task.Status.Status
+			if taskStatus.Status.IsCompleted() {
+				taskStatus.SetEndTime()
+			}
+		}
+		// Update state of the Check
+		newStatus = append(newStatus, taskStatus)
+	}
+
+	for _, ns := range newStatus {
+		summary = common.UpdateStatusSummary(ns.Status, summary)
+	}
+	if common.GetOverallState(summary) != common.StateSucceeded {
+		r.Recorder.Event(workloadInstance, "Warning", "TasksNotFinished", fmt.Sprintf("Tasks have not finished / Namespace: %s, Name: %s, Summary: %v ", workloadInstance.Namespace, workloadInstance.Name, summary))
+	}
+	return newStatus, summary, nil
+}
+
+func GetTaskStatus(taskName string, instanceStatus []klcv1alpha1.TaskStatus) klcv1alpha1.TaskStatus {
+	for _, status := range instanceStatus {
+		if status.TaskDefinitionName == taskName {
+			return status
+		}
+	}
+	return klcv1alpha1.TaskStatus{
+		TaskDefinitionName: taskName,
+		Status:             common.StatePending,
+		TaskName:           "",
+	}
 }
